+++ Spaxel/src/main/java/code/components/sprite/SpriteComponent.java
@@ -11,13 +11,13 @@ import com.fasterxml.jackson.annotation.JsonSetter;
  */
 public class SpriteComponent extends Component {
     private SpriteData sprite;
-    private int scale;
+    private float scale;
 
     private SpriteComponent(){
         super(ComponentType.SPRITE);
     }
 
-    public SpriteComponent(SpriteData sprite, int scale) {
+    public SpriteComponent(SpriteData sprite, float scale) {
         super(ComponentType.SPRITE);
         this.sprite = sprite;
         this.scale = scale;
@@ -36,11 +36,11 @@ public class SpriteComponent extends Component {
         this.sprite = Engine.getEngine().getSpriteAtlas().get(spriteName);
     }
 
-    public int getScale() {
+    public float getScale() {
         return scale;
     }
 
-    public void setScale(int scale) {
+    public void setScale(float scale) {
         this.scale = scale;
     }
 }
